cleanup the edit ingredient form
