+++ meal-ui/src/forms/EditIngredient/index.jsx
@@ -1,0 +1,1 @@
+export { default as IngredientsModalBadge } from './IngredientsModalBadge';
