+++ meal-ui/src/pages/IngredientsIndex/components/index.jsx
@@ -1,2 +1,1 @@
-export { default as IngredientBadge } from './IngredientsBadge';
 export { default as IngredientsIndexRow } from './IngredientsIndexRow';
