--- meal-ui/src/pages/IngredientsIndex/forms/index.jsx
@@ -1,2 +1,0 @@
-export { default as EditIngredientNameForm } from './EditIngredientNameForm';
-export { default as IngredientsModalBadge } from './IngredientsModalBadge';
